Refactor: simplifying the ParserFactory by moving the responsibility of parsing the AST up to the loader.
