+++ modules/extensions-mule-support/src/main/java/org/mule/runtime/module/extension/mule/internal/loader/parser/utils/MuleSdkExtensionLoadingUtils.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.module.extension.mule.internal.loader.parser;
+package org.mule.runtime.module.extension.mule.internal.loader.parser.utils;
 
 import static java.lang.String.format;
 
@@ -13,18 +13,15 @@ import org.mule.runtime.ast.internal.model.ExtensionModelHelper;
 import org.mule.runtime.core.api.type.catalog.ApplicationTypeLoader;
 import org.mule.runtime.extension.api.ExtensionConstants;
 import org.mule.runtime.extension.api.loader.ExtensionLoadingContext;
-import org.mule.runtime.module.extension.internal.loader.parser.ExtensionModelParserFactory;
 
 import java.util.NoSuchElementException;
 
 /**
- * Base {@link ExtensionModelParserFactory} implementation for Mule SDK.
+ * Utilities for loading extensions.
  *
- * Implementations need to specify how to extract some parameters needed by the parser from the extension loading context.
- *
- * @since 4.5.0
+ * @since 4.5
  */
-public abstract class BaseMuleSdkExtensionModelParserFactory implements ExtensionModelParserFactory {
+public class MuleSdkExtensionLoadingUtils {
 
   /**
    * Gets a parameter from the given {@link ExtensionLoadingContext} assuming it needs to be present.
@@ -35,7 +32,7 @@ public abstract class BaseMuleSdkExtensionModelParserFactory implements Extensio
    *
    * @return The parameter value.
    */
-  protected <T> T getRequiredLoadingParameter(ExtensionLoadingContext context, String parameterName) {
+  public static <T> T getRequiredLoadingParameter(ExtensionLoadingContext context, String parameterName) {
     return context.<T>getParameter(parameterName)
         .orElseThrow(() -> new NoSuchElementException(format("Parameter '%s' not found in context", parameterName)));
   }
@@ -44,7 +41,7 @@ public abstract class BaseMuleSdkExtensionModelParserFactory implements Extensio
    * @param context The {@link ExtensionLoadingContext}
    * @return A {@link TypeLoader} suitable for loading types in the context of the extension being modeled in the current context.
    */
-  protected TypeLoader createTypeLoader(ExtensionLoadingContext context) {
+  public static TypeLoader createTypeLoader(ExtensionLoadingContext context) {
     return new ApplicationTypeLoader(context.getDslResolvingContext().getExtensions());
   }
 
@@ -54,7 +51,7 @@ public abstract class BaseMuleSdkExtensionModelParserFactory implements Extensio
    * @return An {@link ExtensionModelHelper} to aid with the lookup of component models by identifier among the extensions in the
    *         context.
    */
-  protected ExtensionModelHelper createExtensionModelHelper(ExtensionLoadingContext context) {
+  public static ExtensionModelHelper createExtensionModelHelper(ExtensionLoadingContext context) {
     return new ExtensionModelHelper(context.getDslResolvingContext().getExtensions(),
                                     context.getDslResolvingContext());
   }
